+++ packages/backend/src/services/user.ts
@@ -1,19 +1,13 @@
 import {
-  AppErrorType,
   GetUserDecoder,
   GetUsersDecoder,
-  handleError,
   NewUserDecoder,
-  PasswordChangeDecoder,
-  PostUserDecoder,
   RawInUserDecoder,
-  validationError,
   type Action,
   type GetUser,
   type Id,
   type NewUser,
   type Optional,
-  type Profile,
   type RawInUser,
   type UserId
 } from "@darkruby/assets-core";
@@ -22,9 +16,8 @@ import { password as Pwd } from "bun";
 import { pipe } from "fp-ts/lib/function";
 import * as TE from "fp-ts/lib/TaskEither";
 import { mapWebError } from "../domain/error";
-import { notFound, type WebAction } from "../fp-express";
+import { type WebAction } from "../fp-express";
 import type { Repository } from "../repository";
-import { verifyPassword } from "./auth";
 
 const parseNewUser = liftTE(NewUserDecoder);
 
@@ -92,67 +85,3 @@ export const deleteUser =
       mapWebError
     );
   };
-
-export const updateProfileOnly =
-  (repo: Repository) =>
-  (id: UserId, body: unknown): WebAction<GetUser> => {
-    return pipe(
-      pipe(body, liftTE(PostUserDecoder)),
-      TE.chain((profile) => repo.user.updateProfileOnly(id, profile)),
-      TE.chain(() => repo.user.get(id)),
-      mapWebError,
-      TE.filterOrElse((u): u is GetUser => Boolean(u), notFound)
-    );
-  };
-
-export const updateOwnProfileOnly =
-  (repo: Repository) =>
-  (userId: UserId, payload: unknown): WebAction<GetUser> =>
-    pipe(
-      TE.Do,
-      TE.bind("ownProfile", () =>
-        pipe(
-          repo.user.get(userId),
-          TE.filterOrElse(
-            (u): u is GetUser => Boolean(u),
-            handleError("Profile not found", AppErrorType.Validation)
-          )
-        )
-      ),
-      TE.bind("profile", () => pipe(payload, liftTE(PostUserDecoder))),
-      mapWebError,
-      TE.chain(({ profile, ownProfile }) =>
-        updateProfileOnly(repo)(userId, {
-          ...ownProfile,
-          username: profile.username
-        })
-      )
-    );
-
-export const updateOwnPasswordOnly =
-  (repo: Repository) =>
-  (profile: Profile, payload: unknown): WebAction<GetUser> =>
-    pipe(
-      TE.Do,
-      TE.tap(() => repo.user.resetAttempts(profile.username)),
-      TE.bind("user", () => repo.user.loginAttempt(profile.username)),
-      TE.bind("passwordChange", () =>
-        pipe(payload, liftTE(PasswordChangeDecoder))
-      ),
-      TE.tap(({ user, passwordChange }) =>
-        pipe(
-          verifyPassword(user.phash, passwordChange.oldPassword),
-          TE.mapLeft(() => validationError("Wrong old password"))
-        )
-      ),
-      TE.chain(({ user, passwordChange }) =>
-        toRawInUser({
-          username: user.username,
-          password: passwordChange.newPassword,
-          admin: user.admin,
-          locked: user.locked
-        })
-      ),
-      TE.chain((user) => repo.user.update(profile.id, user)),
-      mapWebError
-    );
